Substantial models refactor
- use Sequel::Model
  - supports ActiveRecord patterns
    - easier to declare joins and many-to-many association
  - difficult because require presumes the table exists
    already, for dynamic inspection of field properties
    - this is overcome by a trigger when the Models are imported;
      it creates the tables and then requires the models.
  - serialized data support with custom serializers
- removed models from github_repo and github_user
- Models::Manager can persist repos and users to Sqlite, when enabled
