--- spec/fixtures/vcr_cassettes/GithubCrawl_User/_new/with_login/fetches_github_data.yml
@@ -1,77 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.github.com/users/liggitt
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/vnd.github.v3+json
-      User-Agent:
-      - Octokit Ruby Gem 4.8.0
-      Content-Type:
-      - application/json
-      Authorization:
-      - Basic ZGFycmVubGVld2ViZXI6cG90czRnaXRodWI=
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sun, 04 Feb 2018 22:21:10 GMT
-      Content-Type:
-      - application/json; charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Server:
-      - GitHub.com
-      Status:
-      - 200 OK
-      X-Ratelimit-Limit:
-      - '5000'
-      X-Ratelimit-Remaining:
-      - '4949'
-      X-Ratelimit-Reset:
-      - '1517783028'
-      Cache-Control:
-      - private, max-age=60, s-maxage=60
-      Vary:
-      - Accept, Authorization, Cookie, X-GitHub-OTP
-      - Accept-Encoding
-      Etag:
-      - W/"a1f5b84c75e02f3655b09fb4b02441e8"
-      Last-Modified:
-      - Sat, 27 Jan 2018 16:33:26 GMT
-      X-Github-Media-Type:
-      - github.v3; format=json
-      Access-Control-Expose-Headers:
-      - ETag, Link, Retry-After, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining,
-        X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval
-      Access-Control-Allow-Origin:
-      - "*"
-      Content-Security-Policy:
-      - default-src 'none'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubdomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      X-Frame-Options:
-      - deny
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Runtime-Rack:
-      - '0.066930'
-      X-Github-Request-Id:
-      - CFC5:25F91:728A34:992B6E:5A778756
-    body:
-      encoding: ASCII-8BIT
-      string: '{"login":"liggitt","id":980082,"avatar_url":"https://avatars0.githubusercontent.com/u/980082?v=4","gravatar_id":"","url":"https://api.github.com/users/liggitt","html_url":"https://github.com/liggitt","followers_url":"https://api.github.com/users/liggitt/followers","following_url":"https://api.github.com/users/liggitt/following{/other_user}","gists_url":"https://api.github.com/users/liggitt/gists{/gist_id}","starred_url":"https://api.github.com/users/liggitt/starred{/owner}{/repo}","subscriptions_url":"https://api.github.com/users/liggitt/subscriptions","organizations_url":"https://api.github.com/users/liggitt/orgs","repos_url":"https://api.github.com/users/liggitt/repos","events_url":"https://api.github.com/users/liggitt/events{/privacy}","received_events_url":"https://api.github.com/users/liggitt/received_events","type":"User","site_admin":false,"name":"Jordan
-        Liggitt","company":"Red Hat","blog":"","location":"Raleigh, NC","email":"jordan@liggitt.net","hireable":null,"bio":null,"public_repos":105,"public_gists":25,"followers":125,"following":1,"created_at":"2011-08-15T02:08:01Z","updated_at":"2018-01-27T16:33:26Z"}'
-    http_version: 
-  recorded_at: Sun, 04 Feb 2018 22:21:10 GMT
-recorded_with: VCR 4.0.0
